feat: segregate rateService and contract from controller. Validate if chainlink price is 0, check it agains coingeko
