+++ packages/background/src/controllers/ExchangeRatesController.ts
@@ -18,507 +18,13 @@ import { ActionIntervalController } from './block-updates/ActionIntervalControll
 import BlockUpdatesController, {
     BlockUpdatesEvents,
 } from './block-updates/BlockUpdatesController';
-import { BigNumber, Contract } from 'ethers';
-import { formatUnits } from 'ethers/lib/utils';
-
-type DataFeed_Contract = {
-    address: string;
-    abi: any;
-};
-
-type DataFeed_Contracts = {
-    [key: string]: DataFeed_Contract;
-};
-
-const CHAINLINK_DATAFEEDS_CONTRACTS: DataFeed_Contracts = {
-    'ETH/USD': {
-        address: '0x5f4ec3df9cbd43714fe2740f5e3616155c5b8419',
-        abi: [
-            {
-                inputs: [
-                    {
-                        internalType: 'address',
-                        name: '_aggregator',
-                        type: 'address',
-                    },
-                    {
-                        internalType: 'address',
-                        name: '_accessController',
-                        type: 'address',
-                    },
-                ],
-                stateMutability: 'nonpayable',
-                type: 'constructor',
-            },
-            {
-                anonymous: false,
-                inputs: [
-                    {
-                        indexed: true,
-                        internalType: 'int256',
-                        name: 'current',
-                        type: 'int256',
-                    },
-                    {
-                        indexed: true,
-                        internalType: 'uint256',
-                        name: 'roundId',
-                        type: 'uint256',
-                    },
-                    {
-                        indexed: false,
-                        internalType: 'uint256',
-                        name: 'updatedAt',
-                        type: 'uint256',
-                    },
-                ],
-                name: 'AnswerUpdated',
-                type: 'event',
-            },
-            {
-                anonymous: false,
-                inputs: [
-                    {
-                        indexed: true,
-                        internalType: 'uint256',
-                        name: 'roundId',
-                        type: 'uint256',
-                    },
-                    {
-                        indexed: true,
-                        internalType: 'address',
-                        name: 'startedBy',
-                        type: 'address',
-                    },
-                    {
-                        indexed: false,
-                        internalType: 'uint256',
-                        name: 'startedAt',
-                        type: 'uint256',
-                    },
-                ],
-                name: 'NewRound',
-                type: 'event',
-            },
-            {
-                anonymous: false,
-                inputs: [
-                    {
-                        indexed: true,
-                        internalType: 'address',
-                        name: 'from',
-                        type: 'address',
-                    },
-                    {
-                        indexed: true,
-                        internalType: 'address',
-                        name: 'to',
-                        type: 'address',
-                    },
-                ],
-                name: 'OwnershipTransferRequested',
-                type: 'event',
-            },
-            {
-                anonymous: false,
-                inputs: [
-                    {
-                        indexed: true,
-                        internalType: 'address',
-                        name: 'from',
-                        type: 'address',
-                    },
-                    {
-                        indexed: true,
-                        internalType: 'address',
-                        name: 'to',
-                        type: 'address',
-                    },
-                ],
-                name: 'OwnershipTransferred',
-                type: 'event',
-            },
-            {
-                inputs: [],
-                name: 'acceptOwnership',
-                outputs: [],
-                stateMutability: 'nonpayable',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'accessController',
-                outputs: [
-                    {
-                        internalType: 'contract AccessControllerInterface',
-                        name: '',
-                        type: 'address',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'aggregator',
-                outputs: [
-                    { internalType: 'address', name: '', type: 'address' },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'address',
-                        name: '_aggregator',
-                        type: 'address',
-                    },
-                ],
-                name: 'confirmAggregator',
-                outputs: [],
-                stateMutability: 'nonpayable',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'decimals',
-                outputs: [{ internalType: 'uint8', name: '', type: 'uint8' }],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'description',
-                outputs: [{ internalType: 'string', name: '', type: 'string' }],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'uint256',
-                        name: '_roundId',
-                        type: 'uint256',
-                    },
-                ],
-                name: 'getAnswer',
-                outputs: [{ internalType: 'int256', name: '', type: 'int256' }],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'uint80',
-                        name: '_roundId',
-                        type: 'uint80',
-                    },
-                ],
-                name: 'getRoundData',
-                outputs: [
-                    { internalType: 'uint80', name: 'roundId', type: 'uint80' },
-                    { internalType: 'int256', name: 'answer', type: 'int256' },
-                    {
-                        internalType: 'uint256',
-                        name: 'startedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint256',
-                        name: 'updatedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint80',
-                        name: 'answeredInRound',
-                        type: 'uint80',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'uint256',
-                        name: '_roundId',
-                        type: 'uint256',
-                    },
-                ],
-                name: 'getTimestamp',
-                outputs: [
-                    { internalType: 'uint256', name: '', type: 'uint256' },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'latestAnswer',
-                outputs: [{ internalType: 'int256', name: '', type: 'int256' }],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'latestRound',
-                outputs: [
-                    { internalType: 'uint256', name: '', type: 'uint256' },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'latestRoundData',
-                outputs: [
-                    { internalType: 'uint80', name: 'roundId', type: 'uint80' },
-                    { internalType: 'int256', name: 'answer', type: 'int256' },
-                    {
-                        internalType: 'uint256',
-                        name: 'startedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint256',
-                        name: 'updatedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint80',
-                        name: 'answeredInRound',
-                        type: 'uint80',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'latestTimestamp',
-                outputs: [
-                    { internalType: 'uint256', name: '', type: 'uint256' },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'owner',
-                outputs: [
-                    {
-                        internalType: 'address payable',
-                        name: '',
-                        type: 'address',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [{ internalType: 'uint16', name: '', type: 'uint16' }],
-                name: 'phaseAggregators',
-                outputs: [
-                    {
-                        internalType: 'contract AggregatorV2V3Interface',
-                        name: '',
-                        type: 'address',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'phaseId',
-                outputs: [{ internalType: 'uint16', name: '', type: 'uint16' }],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'address',
-                        name: '_aggregator',
-                        type: 'address',
-                    },
-                ],
-                name: 'proposeAggregator',
-                outputs: [],
-                stateMutability: 'nonpayable',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'proposedAggregator',
-                outputs: [
-                    {
-                        internalType: 'contract AggregatorV2V3Interface',
-                        name: '',
-                        type: 'address',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'uint80',
-                        name: '_roundId',
-                        type: 'uint80',
-                    },
-                ],
-                name: 'proposedGetRoundData',
-                outputs: [
-                    { internalType: 'uint80', name: 'roundId', type: 'uint80' },
-                    { internalType: 'int256', name: 'answer', type: 'int256' },
-                    {
-                        internalType: 'uint256',
-                        name: 'startedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint256',
-                        name: 'updatedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint80',
-                        name: 'answeredInRound',
-                        type: 'uint80',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'proposedLatestRoundData',
-                outputs: [
-                    { internalType: 'uint80', name: 'roundId', type: 'uint80' },
-                    { internalType: 'int256', name: 'answer', type: 'int256' },
-                    {
-                        internalType: 'uint256',
-                        name: 'startedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint256',
-                        name: 'updatedAt',
-                        type: 'uint256',
-                    },
-                    {
-                        internalType: 'uint80',
-                        name: 'answeredInRound',
-                        type: 'uint80',
-                    },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    {
-                        internalType: 'address',
-                        name: '_accessController',
-                        type: 'address',
-                    },
-                ],
-                name: 'setController',
-                outputs: [],
-                stateMutability: 'nonpayable',
-                type: 'function',
-            },
-            {
-                inputs: [
-                    { internalType: 'address', name: '_to', type: 'address' },
-                ],
-                name: 'transferOwnership',
-                outputs: [],
-                stateMutability: 'nonpayable',
-                type: 'function',
-            },
-            {
-                inputs: [],
-                name: 'version',
-                outputs: [
-                    { internalType: 'uint256', name: '', type: 'uint256' },
-                ],
-                stateMutability: 'view',
-                type: 'function',
-            },
-        ],
-    },
-};
-
-interface RateService {
-    getRate(
-        currency: string,
-        symbol: string,
-        networkProvider?: any
-    ): Promise<number>;
-}
-
-const chainLinkService: RateService = {
-    async getRate(currency, symbol, networkProvider) {
-        const dataFeedPair =
-            symbol.toUpperCase() + '/' + currency.toUpperCase();
-        const dataFeedAddress =
-            CHAINLINK_DATAFEEDS_CONTRACTS[dataFeedPair]?.address;
-        const dataFeedABI = CHAINLINK_DATAFEEDS_CONTRACTS[dataFeedPair]?.abi;
-
-        if (dataFeedAddress && dataFeedABI) {
-            const chainLinkDataFeedContract = new Contract(
-                dataFeedAddress,
-                dataFeedABI,
-                networkProvider
-            );
-
-            const roundData = await chainLinkDataFeedContract.latestRoundData();
-            return parseFloat(formatUnits(BigNumber.from(roundData.answer), 8));
-        }
-
-        //At this point there is no dataFeedPair in CHAINLINK_DATAFEEDS_CONTRACTS
-        return 0;
-    },
-};
-
-const coingekoService: RateService = {
-    async getRate(currency, symbol) {
-        const baseApiEndpoint = 'https://api.coingecko.com/api/v3/simple/';
-        const query = `${baseApiEndpoint}price`;
-        const currencyApiId =
-            RATES_IDS_LIST[symbol.toUpperCase() as keyof typeof RATES_IDS_LIST];
-
-        const response = await axios.get(query, {
-            params: {
-                ids: currencyApiId,
-                vs_currencies: currency,
-            },
-        });
-
-        if (response.status != 200) {
-            throw new Error(response.statusText);
-        }
-
-        return response.data[currencyApiId][currency];
-    },
-};
-
-const rateProvider: Record<string, RateService> = {
-    'usd-eth': chainLinkService,
-};
-
-const getRateService = (
-    nativeCurrency: string,
-    tokenSymbol: string
-): RateService => {
-    const provider =
-        rateProvider[`${nativeCurrency}-${tokenSymbol}`.toLowerCase()];
-    return provider || coingekoService;
-};
+import {
+    getRateService,
+    RateService,
+    BaseApiEndpoint,
+    chainLinkService,
+    coingekoService,
+} from '../utils/rateService';
 
 export interface ExchangeRatesControllerState {
     exchangeRates: Rates;
@@ -534,8 +40,6 @@ export interface Rates {
 
 const DEFAULT_NETWORK_CURRENCY_ID = 'ethereum';
 export class ExchangeRatesController extends BaseController<ExchangeRatesControllerState> {
-    private readonly baseApiEndpoint =
-        'https://api.coingecko.com/api/v3/simple/';
     private readonly staticTokens: { [address: string]: { token: Token } } = {
         '0x6B175474E89094C44Da98b954EedeAC495271d0F': {
             token: { symbol: 'DAI' } as Token,
@@ -660,6 +164,17 @@ export class ExchangeRatesController extends BaseController<ExchangeRatesControl
             // this._networkController.getProvider()
         );
 
+        //In case chainlink returns 0, we will retrieve value from Coingeko, in case of failure from chainlink
+        if (
+            rates[symbol] === 0 &&
+            this._exchangeRateService === chainLinkService
+        ) {
+            rates[symbol] = await coingekoService.getRate(
+                nativeCurrency,
+                symbol
+            );
+        }
+
         // Get tokens exchange rates
         const tokenRatesQuery = await this._getTokenRates();
 
@@ -688,7 +203,7 @@ export class ExchangeRatesController extends BaseController<ExchangeRatesControl
         const tokens = { ...this.staticTokens, ...this.getTokens() };
         const tokenContracts = Object.keys(tokens).join(',');
 
-        const query = `${this.baseApiEndpoint}token_price/${this.networkNativeCurrency.coingeckoPlatformId}`;
+        const query = `${BaseApiEndpoint}token_price/${this.networkNativeCurrency.coingeckoPlatformId}`;
 
         const response = await axios.get(query, {
             params: {
